get screen size and clearly XY location

+++ app/src/main/java/com/example/escapegame/East.java
@@ -43,8 +43,8 @@ public class East extends AppCompatActivity {
 
     public boolean onTouchEvent(MotionEvent motionEvent) {
 
-        float xplace = motionEvent.getX();
-        float yplace = motionEvent.getY();
+        int xplace = (int) motionEvent.getX();
+        int yplace = (int) motionEvent.getY();
 
         switch (motionEvent.getAction()) {
