+++ app/src/main/java/com/example/escapegame/South.java
@@ -42,8 +42,8 @@ public class South extends AppCompatActivity {
 
     public boolean onTouchEvent(MotionEvent motionEvent) {
 
-        float xplace = motionEvent.getX();
-        float yplace = motionEvent.getY();
+        int xplace = (int) motionEvent.getX();
+        int yplace = (int) motionEvent.getY();
 
         switch (motionEvent.getAction()) {
